// src/app/api/models/user-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserResponse {
  createdAt: string;
  deletedAt: string | null;
  email: string;
  emailVerifiedAt: string | null;
  id: string;
  lastName: string | null;
  name: string;
  passwordConfirmedAt: string | null;
  permissions: Array<{
'action': string;
'subject': string;
}>;
  phone: string | null;
  roles: Array<{
'id': string;
'name': string;
'description': string | null;
}>;
  updatedAt: string;
  username: string;
  dateOfBirth: string | null;
  address: string | null;
  address2: string | null;
  zipCode: string | null;
  city: string | null;
  state: string | null;
  country: string | null;
  gender: string | null;
  profilePhotoUrl: string | null;
  totpEnabled: boolean;
  mustChangePassword: boolean;
}